refactor(google-maps): remove unnecessary modules
Removes the `GoogleMapModule` and `MapMarkerModule` since they're redundant.
