--- src/google-maps/google-map/google-map-module.ts
@@ -1,18 +1,0 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
-import {NgModule} from '@angular/core';
-
-import {GoogleMap} from './google-map';
-
-@NgModule({
-  exports: [GoogleMap],
-  declarations: [GoogleMap],
-})
-export class GoogleMapModule {
-}
