--- src/google-maps/map-info-window/index.ts
@@ -1,10 +1,0 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
-export * from './map-info-window';
-export * from './map-info-window-module';
