+++ src/google-maps/map-info-window/map-info-window.spec.ts
@@ -2,8 +2,8 @@ import {Component} from '@angular/core';
 import {async, TestBed} from '@angular/core/testing';
 import {By} from '@angular/platform-browser';
 
-import {DEFAULT_OPTIONS, GoogleMapModule, UpdatedGoogleMap} from '../google-map/index';
-import {MapMarker} from '../map-marker/index';
+import {DEFAULT_OPTIONS, UpdatedGoogleMap} from '../google-map/google-map';
+import {MapMarker} from '../map-marker/map-marker';
 import {
   createInfoWindowConstructorSpy,
   createInfoWindowSpy,
@@ -12,17 +12,15 @@ import {
   TestingWindow
 } from '../testing/fake-google-map-utils';
 
-import {MapInfoWindow, MapInfoWindowModule} from './index';
+import {GoogleMapsModule} from '../google-maps-module';
+import {MapInfoWindow} from './map-info-window';
 
 describe('MapInfoWindow', () => {
   let mapSpy: jasmine.SpyObj<UpdatedGoogleMap>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [
-        GoogleMapModule,
-        MapInfoWindowModule,
-      ],
+      imports: [GoogleMapsModule],
       declarations: [TestApp],
     });
   }));
