--- src/google-maps/map-marker/map-marker-module.ts
@@ -1,17 +1,0 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
-import {NgModule} from '@angular/core';
-import {MapMarker} from './map-marker';
-
-@NgModule({
-  exports: [MapMarker],
-  declarations: [MapMarker],
-})
-export class MapMarkerModule {
-}
