+++ src/google-maps/public-api.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-export {GoogleMap} from './google-map/index';
-export {MapInfoWindow} from './map-info-window/index';
-export {MapMarker} from './map-marker/index';
-export * from './google-maps-module';
+export {GoogleMap} from './google-map/google-map';
+export {MapMarker} from './map-marker/map-marker';
+export {GoogleMapsModule} from './google-maps-module';
+export {MapInfoWindow} from './map-info-window/map-info-window';
